+++ npe2/__init__.py
@@ -6,13 +6,13 @@ __author__ = "Talley Lambert"
 __email__ = "talley.lambert@gmail.com"
 
 from ._plugin_manager import PluginContext, PluginManager
-from .io_utils import read, write_layer_data
+from .io_utils import read, write
 from .manifest import PluginManifest
 
 __all__ = [
     "PluginManifest",
     "PluginManager",
     "PluginContext",
-    "write_layer_data",
+    "write",
     "read",
 ]
