+++ tests/test_npe2.py
@@ -4,8 +4,7 @@ import sys
 import pytest
 from pydantic import ValidationError
 
-from npe2 import PluginManager, PluginManifest, write_layer_data
-from npe2._types import FullLayerData
+from npe2 import PluginManager, PluginManifest
 from npe2.cli import main
 
 
@@ -243,33 +242,3 @@ def test_writer_valid_layer_type_expressions(expr, uses_sample_plugin):
     data["contributions"]["writers"][0]["layer_types"].append(expr)
 
     PluginManifest(**data)
-
-
-null_image: FullLayerData = ([], {}, "image")
-
-
-def test_writer_exec(uses_sample_plugin):
-    # the sample writer knows how to handle two image layers
-    result = write_layer_data("test.tif", [null_image, null_image])
-    assert result == ["test.tif"]
-
-
-@pytest.mark.parametrize("layer_data", [[null_image, null_image], []])
-def test_writer_exec_fails(layer_data, uses_sample_plugin):
-    # the sample writer doesn't accept no extension
-    with pytest.raises(ValueError):
-        write_layer_data("test/path", layer_data)
-
-
-def test_writer_exec_fails2(uses_sample_plugin):
-    # the sample writer doesn't accept 5 images
-    with pytest.raises(ValueError):
-        write_layer_data(
-            "test.tif", [null_image, null_image, null_image, null_image, null_image]
-        )
-
-
-def test_writer_single_layer_api_exec(uses_sample_plugin):
-    # This writer doesn't do anything but type check.
-    paths = write_layer_data("test/path", [([], {}, "labels")])
-    assert len(paths) == 1
